fix: dark mode: format and example border now same color as section border
in dark mode, the format and example box has a small border to distinguish itself from the rest of the section. This is not present in light mode. It used to be a slate-600 color, it is now the same color as the border around the entire section, black. In addition, the size has been changed from border-2 to -1 to be consistant with the section border.